+++ config/routes.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 RecordingStudioUser::Engine.routes.draw do
-  resource :profile, only: %i[show edit update], path: RecordingStudioUser.config.profile_route_path
+  resource :profile, only: %i[show edit update], path: RecordingStudioUser.config.profile_route_path do
+    patch :photo, action: :update_photo
+  end
   get RecordingStudioUser.config.admin_route_path, to: "admin/users#index", as: :admin
 end