+++ test/engine_test.rb
@@ -14,6 +14,7 @@ class EngineTest < Minitest::Test
 
     assert_includes routes, "profile"
     assert_includes routes, "edit_profile"
+    assert_includes routes, "photo_profile"
     assert_includes routes, "admin"
     refute_includes routes, "admin_user"
     refute_includes routes, "edit_admin_user"